feat: restructure dashboard layout and integrate file manager component
