+++ app/(navigation)/dashboard/forms/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function page() {
+  return (
+    <>
+      <div>Intake Form Page</div>
+    </>
+  )
+}
+
+export default page
