+++ app/(navigation)/dashboard/layout.tsx
@@ -1,0 +1,24 @@
+import SideBar from '@/components/ui/sideBar';
+import React from 'react';
+
+interface DashboardLayoutProps {
+  children: React.ReactNode;
+}
+
+export default function DashboardLayout({ children }: DashboardLayoutProps) {
+  return (
+    <div className="flex h-screen">
+      {/* Sidebar */}
+      <div className="flex-shrink-0">
+        <SideBar />
+      </div>
+      
+      {/* Main content area */}
+      <div className="flex-1 overflow-auto bg-white">
+        <div className="p-6">
+          {children}
+        </div>
+      </div>
+    </div>
+  );
+}
