+++ app/(navigation)/dashboard/messages/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function page() {
+  return (
+    <>
+    Messages coming soon
+    </>
+  )
+}
+
+export default page
