+++ app/(navigation)/dashboard/page.tsx
@@ -1,13 +1,13 @@
 
-import SideBar from '@/components/ui/sideBar'
-import { ApptTable } from '@/components/ui/ApptTable'
 import React from 'react'
-
 
 function page() {
   return (
-    <>
-    <SideBar />
-    <ApptTable />
-    </>
+    <div>
+      <h1 className="text-2xl font-bold mb-6">Dashboard Overview</h1>
+      <p>Welcome to your dashboard. Use the sidebar to navigate to different sections.</p>
+    </div>
   )
+}
+
+export default page;
